requirements.txt
asgiref
certifi
click
colorama
crispy-bootstrap5
distlib
dj-database-url
Django
django-allauth
django-braces
django-cors-headers
django-crispy-forms
django-environ
django-filter
django-humanize
django-model-utils
django-star-ratings
djangorestframework
filelock
gunicorn
h11
humanize
Markdown
MouseInfo
numpy
opencv-contrib-python
packaging
pillow
pipenv
platformdirs
psycopg2
psycopg2-binary
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
PyScreeze
pytweening
setuptools
sqlparse
swapper
typing_extensions
tzdata
uvicorn
virtualenv